Update and rename README.md to 2020 - The Dawn of A New Era
My story starts in 2012 when my Dad had a neck (postural) surgery from the C3-C7 using titanium rods, plates, and screws. After being discharged from the hospital he was prescribed opioid medication to manage the pain. In February 2013, my Dad needed a revision surgery on his neck due to complications and continued to use opioid medication to manage his chronic pain. While he was recovering, he tore his meniscus in his left knee and required surgery in May 2013. In 2015, my Dad had an assessment at Allevio Pain Management Clinic with Dr. Kevin Smith and was introduced to the ketamine-lidocaine infusion. Eight months later, my Dad was no longer taking opioids as the pain was managed through this treatment every two months. My Dad continues to receive this treatment with Dr. Kevin Smith at the Silver Medical Group today and remains off opioids. In December 2016, my Dad had a lumber surgery and ten days later he had developed severe head pain in the postural position. A spinal fluid leak (CFS leak) was detected and required another surgery to repair the leak in January 2017. Once again opioids were the only medication prescribed to manage his chronic pain. In April 2017 my Dad was on his way to work when he was rear-ended by an SUV on highway 400. He was diagnosed with a spinal cord concussion which led to PTSD, anxiety, and depression, in addition to his chronic pain.


Medical cannabis has significantly improved my Dad’s quality of life and, as a result, the rest of my family’s quality of life. An opportunity presented itself to transition into this emerging industry by helping my Dad as a registered patient under the Access to Cannabis for Medical Purposes Regulations (ACMPR) to understand the legal framework. The journey was not easy and disappointment was at every corner. Nonetheless, we did not let regulatory issues, further delays in production, people and money stop us. We remained focused and sought after a partner that operates for the local medical community and is backed up by science. It is important for us to work with a medical Licensed Producer that is committed to advancing science and research through clinical studies that the industry requires to reach its full potential. While we were interested in Green Relief’s philosophy, we knew we had to tour the facility to meet the team and see for ourselves if we had similar synergies. 


I was walking through the facility and my expectations were exceeded. In my opinion, Green Relief truly does represent the gold standard for medical cannabis. As a Canadian, I feel pride witnessing what I thought was the most incredible state-of-the-art facility and one of its kind, right here in Ontario. We stand by Green Relief because of the biotechnological research that they undertake. Their medical patient-focussed program involves more time spent with patients providing them with the support and care they need, as well as constantly improving their medicine. The CEO, Dr. Neilank K. Jha is a Neurosurgeon and International Leading Expert in Traumatic Brain Injury.[1] Dr. Jha is a major reason for which we are motivated to move forward with Green Relief as he would be perfect to collaborate with on the issue of chronic pain. In addition to hearing that Dr. Jha has trained and worked with my Dad’s neurosurgeon, Dr. Michael Fehlings at Toronto Western Hospital.[2] I was pleasantly surprised to hear the neurosurgeons history and I became incredibly excited as their background is critical for breakthrough medicine for those with chronic pain as a result of spinal surgery. The technology Green Relief is using to extract cannabinoid compounds is extraordinary and currently are creating the purest quality of CBD in the market today.

The high level of organization and cleanliness was impeccable as we were walking through the corridors and into each room. Green Relief exercises the highest standards of hygiene and follows Health Canada’s guidelines so much so that Health Canada utilizes Green Relief’s facility for their own training purposes. I strongly believe in a brand that fulfills its social responsibly and for Green Relief, they are donating the fish used in their aquaponic process to a local shelter each month to feed those in need, in addition to exploring ways to integrate job positions for Canadians through the Ontario Disability Employment Network.[3]

I was fascinated to learn about Green Relief’s aquaponic process for medical cannabis and I have to admit, I was quite skeptical about how the system actually operated. Aquaponics blend aquaculture and hydroponics together and is the most sustainable form of agriculture: it uses 90% less water than conventional farming and highly efficient for developing effective medicine.[4] The future for sustainable agriculture is through Green Relief’s highly balanced ecosystem and its innovative technology. Together it combines to create a remarkable pillar that will help propel the medical cannabis industry. I would like to give a special thank you to Steve LeBlanc who we toured the facility with and the team at Green Relief for their hospitality.

[1] https://www.greenrelief.ca/team/

[2] https://www.uhn.ca/PatientsFamilies/Search_Doctors/Pages/doctor_detail.aspx?doctor=361

[3] https://www.greenrelief.ca/

[4] https://www.greenrelief.ca/